fix: Handle dtype conversion for MLX-PyTorch compatibility
MLX returns uint32 for integer arrays, but MPS backend doesn't
support this type. Convert via CPU first, then move to MPS.

Changes:
- Convert MLX arrays to CPU first
- Apply .long() on CPU (uint32 → int64)
- Then move to MPS device
- Keeps everything in native MLX until conversion

This fixes: TypeError: Trying to convert UInt32 to MPS backend

Pure MLX solution - no fallback to PyTorch needed.

+++ indextts/gpt/mlx_model.py
@@ -229,12 +229,10 @@ class UnifiedVoiceMLX(nn.Module):
             max_length=kwargs.get('max_generate_length', 1500)
         )
         
-        # Convert back to PyTorch (to CPU first to handle dtype conversion)
-        codes = mlx_to_torch(codes_mlx, device='cpu')
-        conditioning_latent = mlx_to_torch(conditioning, device='cpu')
-        
-        # Ensure correct dtype (int64/long) before moving to MPS
-        codes = codes.long()
+        # Convert MLX arrays to PyTorch via CPU to handle dtype properly
+        # MLX returns uint32, need int64 for PyTorch
+        codes = mlx_to_torch(codes_mlx, device='cpu').long()  # Convert to int64 on CPU
+        conditioning_latent = mlx_to_torch(conditioning, device='cpu')  # Float32 is fine
         
         # Now move to MPS device
         codes = codes.to('mps')
